app/requirements.txt
flask
flask-dance[sqla]
blinker
python-dotenv